Move examples to examples folder
